Made the board way more awesome!!!
Swapped from ImageIcon to painting a background image on the JPanel for each BoardSpace.

Modified player icons and containing grid to be smaller and fit 8 side by side and well.

Added house and hotel pngs and formatted their images all ready to be put on the board.

Added a new GridPanel to hold houses/hotel for each BoardSpace.
 - Evan